Lcg + Client

+++ Lab3/Crackers/LcgCracker.cs
@@ -9,19 +9,19 @@ namespace Lab3
 {
     class LcgCracker
     {
-        private BigInteger a;
-        private BigInteger c;
-        public static readonly BigInteger m = 4294967296;
-        private BigInteger _last;
+        private long a;
+        private long c;
+        public static readonly long m = 4294967296;
+        private long _last;
 
-        public LcgCracker(BigInteger a, BigInteger c, BigInteger last)
+        public LcgCracker(long a, long c, long last)
         {
             this.a = a;
             this.c = c;
             _last = last;
         }
 
-        public BigInteger Next()
+        public long Next()
         {
             _last = (a * _last + c) % m;
             return _last;
